refactor: lastModified & dateToString 프로퍼티수정
